refactor(editor): modulariza configuracao de widgets e filtros
